feat(*): add addToCart and permissions

+++ backend/src/resolvers/Mutation.js
@@ -3,19 +3,17 @@ const jwt = require('jsonwebtoken')
 const { randomBytes } = require('crypto')
 const { promisify } = require('util')
 const { transport, makeANiceEmail } = require('../mail')
+const {hasPermission} = require('./../utils')
 
 const Mutations = {
     async createItem(parent, args, ctx, info) {
-        if(!ctx.request.userID) {
-            throw new Error('you must be logged in to do that')
-        }
         const item = await ctx.db.mutation.createItem({
             data: {
                 user: {
                     connect: {
                         id: ctx.request.userId
                     }
-                }
+                },
                 ...args
             }
         }, info)
@@ -36,9 +34,23 @@ const Mutations = {
         )
     },
     async deleteItem(parent, args, ctx, info) {
+
         const where = { id: args.id };
 
-        const item = await ctx.db.query.item({where}, `{ id title}`)
+        const item = await ctx.db.query.item({where}, `{ 
+            id 
+            title 
+            user {
+                id
+            } 
+        }`)
+
+        const ownsItem = item.user.id === ctx.request.userId
+        const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission))
+
+        if(!ownsItem && !hasPermissions ) {
+            throw new Error('you dont have permissions to do that')
+        }
 
         return ctx.db.mutation.deleteItem({
             where
@@ -56,7 +68,7 @@ const Mutations = {
             data: {
                 ...args,
                 password,
-                permissions: { set: ['User']}
+                permissions: { set: ['USER']}
             }
         }, info)
 
@@ -164,6 +176,73 @@ const Mutations = {
         });
         // 8. return the new user
         return updatedUser;
+    },
+
+    async updatePermissions(parent, args, ctx, info) {
+        // login check
+        if(!ctx.request.userId) {
+            throw new Error('you must be logged in to do that')
+        }
+
+        // query the user
+
+        const currentUser = await ctx.db.query.user({
+            where: { id: ctx.request.userId}
+        }, info)
+
+        
+
+        // check if permissions
+        hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE'])
+
+        // update
+        return ctx.db.mutation.updateUser({
+            data: {
+                permissions: {
+                    set: args.permissions
+                }
+            },
+            where: {
+                id: args.userId
+            },
+        }, info)
+    },
+
+    async addToCart(parent, args, ctx, info) {
+        const {userId} = ctx.request
+        // check if signed in 
+        if(!userId) {
+            throw new Error('you must be logged in to do that')
+        }
+
+        // query user cart item
+        const [existingCartItem] = await ctx.db.query.cartItems({
+            where: {
+                user: {id: userId},
+                item: {id: args.id}
+            }
+        })
+
+        // check if already in cart
+        if(existingCartItem) {
+            console.log('already present in cart')
+            return ctx.db.mutation.updateCartItem({
+                where: { id: existingCartItem.id},
+                data: {quantity: existingCartItem.quantity + 1}
+            }, info)
+        }
+
+        // if its not create new connection
+        return ctx.db.mutation.createCartItem({
+            data: {
+                user: {
+                    connect: {id: userId},
+                },
+                item: {
+                    connect: {id: args.id}
+                } 
+            }
+        }, info)
     }
 };
